+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 next
 ----
 - Add `Parameters::next_array` function returning placeholder as an array, intended to be used with INSERT INTO queries
+- Rename `WithClause::and` to `and_with` for clarity
 
 0.3.0
 -----